checkpoint: moves ".../Fetching/ResponseError.ts" to ".../Fetching/Error/Response.ts"
